+++ examples/configurations/dedicated-transit-gateway-attachment-aws-kafka-rbac/main.tf
@@ -3,7 +3,7 @@ terraform {
   required_providers {
     confluent = {
       source  = "confluentinc/confluent"
-      version = "1.32.0"
+      version = "1.36.0"
     }
     aws = {
       source  = "hashicorp/aws"