+++ examples/configurations/ksql-rbac/main.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     confluent = {
       source  = "confluentinc/confluent"
-      version = "1.32.0"
+      version = "1.36.0"
     }
   }
 }
@@ -39,12 +39,12 @@ resource "confluent_schema_registry_cluster" "essentials" {
 
 # Update the config to use a cloud provider and region of your choice.
 # https://registry.terraform.io/providers/confluentinc/confluent/latest/docs/resources/confluent_kafka_cluster
-resource "confluent_kafka_cluster" "basic" {
+resource "confluent_kafka_cluster" "standard" {
   display_name = "inventory"
   availability = "SINGLE_ZONE"
   cloud        = "AWS"
   region       = "us-east-2"
-  basic {}
+  standard {}
   environment {
     id = confluent_environment.staging.id
   }
@@ -59,7 +59,7 @@ resource "confluent_service_account" "app-manager" {
 resource "confluent_role_binding" "app-manager-kafka-cluster-admin" {
   principal   = "User:${confluent_service_account.app-manager.id}"
   role_name   = "CloudClusterAdmin"
-  crn_pattern = confluent_kafka_cluster.basic.rbac_crn
+  crn_pattern = confluent_kafka_cluster.standard.rbac_crn
 }
 
 resource "confluent_api_key" "app-manager-kafka-api-key" {
@@ -72,9 +72,9 @@ resource "confluent_api_key" "app-manager-kafka-api-key" {
   }
 
   managed_resource {
-    id          = confluent_kafka_cluster.basic.id
-    api_version = confluent_kafka_cluster.basic.api_version
-    kind        = confluent_kafka_cluster.basic.kind
+    id          = confluent_kafka_cluster.standard.id
+    api_version = confluent_kafka_cluster.standard.api_version
+    kind        = confluent_kafka_cluster.standard.kind
 
     environment {
       id = confluent_environment.staging.id
@@ -95,26 +95,46 @@ resource "confluent_api_key" "app-manager-kafka-api-key" {
 
 resource "confluent_kafka_topic" "users" {
   kafka_cluster {
-    id = confluent_kafka_cluster.basic.id
+    id = confluent_kafka_cluster.standard.id
   }
   topic_name    = "users"
-  rest_endpoint = confluent_kafka_cluster.basic.rest_endpoint
+  rest_endpoint = confluent_kafka_cluster.standard.rest_endpoint
   credentials {
     key    = confluent_api_key.app-manager-kafka-api-key.id
     secret = confluent_api_key.app-manager-kafka-api-key.secret
   }
 }
 
-// Service account that the ksqlDB cluster uses to talk to the Kakfa broker.
+// ksqlDB service account
 resource "confluent_service_account" "app-ksql" {
   display_name = "app-ksql"
-  description  = "Service account to manage 'inventory' ksqlDB cluster"
+  description  = "Service account for ksqlDB cluster"
 }
 
-resource "confluent_role_binding" "app-ksql-kafka-cluster-admin" {
+resource "confluent_role_binding" "app-ksql-all-topic" {
   principal   = "User:${confluent_service_account.app-ksql.id}"
-  role_name   = "CloudClusterAdmin"
-  crn_pattern = confluent_kafka_cluster.basic.rbac_crn
+  role_name   = "ResourceOwner"
+  crn_pattern = "${confluent_kafka_cluster.standard.rbac_crn}/kafka=${confluent_kafka_cluster.standard.id}/topic=*"
+}
+
+resource "confluent_role_binding" "app-ksql-all-group" {
+  principal   = "User:${confluent_service_account.app-ksql.id}"
+  role_name   = "ResourceOwner"
+  crn_pattern = "${confluent_kafka_cluster.standard.rbac_crn}/kafka=${confluent_kafka_cluster.standard.id}/group=*"
+}
+
+resource "confluent_role_binding" "app-ksql-all-transactions" {
+  principal   = "User:${confluent_service_account.app-ksql.id}"
+  role_name   = "ResourceOwner"
+  crn_pattern = "${confluent_kafka_cluster.standard.rbac_crn}/kafka=${confluent_kafka_cluster.standard.id}/transactional-id=*"
+}
+
+# ResourceOwner roles above are for KSQL service account to read/write data from/to kafka,
+# this role instead is needed for giving access to the Ksql cluster.
+resource "confluent_role_binding" "app-ksql-ksql-admin" {
+  principal   = "User:${confluent_service_account.app-ksql.id}"
+  role_name   = "KsqlAdmin"
+  crn_pattern = confluent_ksql_cluster.main.resource_name
 }
 
 resource "confluent_role_binding" "app-ksql-schema-registry-resource-owner" {
@@ -127,7 +147,7 @@ resource "confluent_ksql_cluster" "main" {
   display_name = "ksql_cluster_0"
   csu = 1
   kafka_cluster {
-    id = confluent_kafka_cluster.basic.id
+    id = confluent_kafka_cluster.standard.id
   }
   credential_identity {
     id = confluent_service_account.app-ksql.id
@@ -136,7 +156,6 @@ resource "confluent_ksql_cluster" "main" {
     id = confluent_environment.staging.id
   }
   depends_on = [
-    confluent_role_binding.app-ksql-kafka-cluster-admin,
     confluent_role_binding.app-ksql-schema-registry-resource-owner,
     confluent_schema_registry_cluster.essentials
   ]